video loading msgs refactor
