+++ client/src/components/Video/Message/Message.css
@@ -1,0 +1,17 @@
+.Video__Wrapper--Message{
+    margin-left:auto; 
+    margin-right:auto;
+    width: 100%;
+    position: relative;
+}
+
+.Video--Message{
+    position: fixed;
+    top: 25vh;
+    left: 25vw;
+    width: 60vw;
+    height: 60vh;
+    max-width:640px;
+    color:white;
+    text-align:center;  
+}
