+++ client/src/components/Video/Video.css
@@ -19,21 +19,5 @@
     max-width:640px;
 }
 
-.Video__Wrapper--Message{
-    margin-left:auto; 
-    margin-right:auto;
-    width: 100%;
-    position: relative;
-}
 
-.Video--Message{
-    position: fixed;
-    top: 25vh;
-    left: 25vw;
-    width: 60vw;
-    height: 60vh;
-    max-width:640px;
-    color:white;
-    text-align:center;  
-}
